+++ api/stateless/lib/hub/remote.ts
@@ -158,6 +158,10 @@ export default class RemoteHub implements HubClient {
         await this.#call('/event/set', { layerid, cron });
     }
 
+    async featureRefresh(connection: number): Promise<void> {
+        await this.#call('/feature/refresh', { connection });
+    }
+
     async coreEventSubmit(event: string): Promise<void> {
         await this.#call(`/core/event/${encodeURIComponent(event)}`, {});
     }